Initial schema for event with test
