+++ .babelrc
@@ -1,4 +1,4 @@
 {
-  "presets": ["es2015", "stage-2"],
+  "presets": ["es2015", "stage-0"],
   "plugins": []
 }
