+++ lib/index.spec.js
@@ -3,7 +3,7 @@ import assert from 'assert';
 
 describe('Example Node Server', () => {
   it('should return 200', done => {
-    http.get('http://127.0.0.1:1337', res => {
+    http.get('http://localhost:3001/graphql', res => {
       assert.equal(200, res.statusCode);
       done();
     });
