+++ package.json
@@ -4,10 +4,10 @@
   "description": "GraphQL server app",
   "main": "index.js",
   "scripts": {
-    "start": "nodemon lib/index.js --exec babel-node --presets es2015,stage-2",
-    "build": "babel lib -d dist --presets es2015,stage-2",
+    "start": "nodemon lib/index.js --exec babel-node --presets es2015,stage-0",
+    "build": "babel lib -d dist --presets es2015,stage-0",
     "serve": "node dist/index.js",
-    "test": "mocha lib/index.spec.js --compilers js:babel-register"
+    "test": "mocha lib/query.spec.js --compilers js:babel-register --require babel-polyfill -R mocha-trumpet-reporter"
   },
   "repository": {
     "type": "git",
@@ -29,9 +29,13 @@
   },
   "devDependencies": {
     "babel-cli": "^6.9.0",
+    "babel-polyfill": "^6.9.1",
     "babel-preset-es2015": "^6.9.0",
+    "babel-preset-stage-0": "^6.5.0",
     "babel-preset-stage-2": "^6.5.0",
     "babel-register": "^6.9.0",
-    "mocha": "^2.5.3"
+    "chai": "^3.5.0",
+    "mocha": "^2.5.3",
+    "mocha-trumpet-reporter": "0.0.3"
   }
 }
